Continue work on automatic user registration
Issue #185

+++ nunaliit2-security/src/main/java/ca/carleton/gcrc/security/ber/BerInteger.java
@@ -2,7 +2,7 @@ package ca.carleton.gcrc.security.ber;
 
 public interface BerInteger extends BerObject {
 
-	public Integer getValue();
-	public void setValue(Integer value);
+	public Long getValue();
+	public void setValue(Long value);
 	
 }
